Move binops, indexing, remove unused series files
