--- python/cudf/cudf/tests/series/test_binary_operations.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
