+++ python/cudf/cudf/tests/series/test_binops.py
@@ -1,0 +1,35 @@
+# Copyright (c) 2025, NVIDIA CORPORATION.
+import operator
+
+import pandas as pd
+import pytest
+
+import cudf
+from cudf.testing._utils import (
+    assert_exceptions_equal,
+)
+
+
+@pytest.mark.parametrize(
+    "sr1", [pd.Series([10, 11, 12], index=["a", "b", "z"]), pd.Series(["a"])]
+)
+@pytest.mark.parametrize(
+    "sr2",
+    [pd.Series([], dtype="float64"), pd.Series(["a", "a", "c", "z", "A"])],
+)
+@pytest.mark.parametrize(
+    "op",
+    [
+        operator.eq,
+        operator.ne,
+        operator.lt,
+        operator.gt,
+        operator.le,
+        operator.ge,
+    ],
+)
+def test_series_error_equality(sr1, sr2, op):
+    gsr1 = cudf.from_pandas(sr1)
+    gsr2 = cudf.from_pandas(sr2)
+
+    assert_exceptions_equal(op, op, ([sr1, sr2],), ([gsr1, gsr2],))
