--- python/cudf/cudf/tests/series/test_function_application.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
