--- python/cudf/cudf/tests/series/test_io_serialization.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
