--- python/cudf/cudf/tests/series/test_missing.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
