--- python/cudf/cudf/tests/series/test_selecting.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
