--- python/cudf/cudf/tests/series/test_timeseries.py
@@ -1,1 +1,0 @@
-# Copyright (c) 2023, NVIDIA CORPORATION.
